tests/winetricks-test: fix running tests without xvfb
This was broken in 086155bdf6fb43321364ca8ac8a8f04bc673630e.

The unnecessary quoting caused winetricks-tests to run winetricks like:
$ "" sh src/winetricks --no-isolate -q --verify foo

if not running xvfb tests. This results in:
bash: : command not found

Removing the quotes works around the problem, and is safe as $XVFB is only
defined to one possible known value ('xvfb-run') that does not need quoting.