verbs: CSMT enabled by default
Since Wine 3.3, CSMT is enabled by default (see https://www.winehq.org//announce/3.3).
However, the setting options CSMT is by default disabled according to
the Winetricks settings menu. This commit updates the titles for these
options to the correct ones. It also swaps the order of the CSMT options,
since in most cases the default option is the top setting.