Swap Tahoma version v2.26 for v2.60
Put simply the core fonts project have had the IELPKTH.CAB file laying around on their sourceforge account for ages, most Linux users when needing Tahoma probably find it (or a similar copy of IELPKTH.CAB) there first.

I have no rationale for wanting version 2.60 over 2.26 other than to say that if we were not interested in newer versioned fonts, then why does verb 'eufonts' exist when that's all that does?

Looking through FontForge the only thing I can see different between these versions is 10 more symbols... however both of them already have the euro symbol (so if that is the reason for the 'eufonts' verb to exist is for adding the euro symbol, that reason isn't applicable here).

Also it's worth noting that although the file currently being downloaded is called "tahoma32.exe" which looks similarly named to the other  with something32.exe named files from the 'Core fonts for the Web' effort... that naming is coincidental.  Tahoma is not a part of the 'Core fonts for the Web'.