winetricks_set_wineprefix: remove unneeded check for 'unknown' string
This was originally added in fcb79b70ea38653a69ef0e0ae2291815ee69ec7e,
though I can't find the reason or the source in wine where that happens
(in 1.3.14, which was current at the time, or 5.7).

Given that it's breaking users with 'unknown' for a username, and that
if the WINEPREFIX really is borked, things will fail later anyway, I
think it's safe to remove until we have a better idea why it's needed.