w_download_to: checksum ALL files
I.e., remove the exclusion for files > 500MB.

This was originally intendeded (AFAIR) to exclude the EVE online
installer.

Since then, it has allowed the win7sp1 installer (et al) to break, as
they are over the size limit, but if users get a corrupted version (that
is > 500MB), winetricks will still attempt to install.

Closes: #1916