Fix remove_mono breaks dotnet40 on subsequent dotnet* installs
We should not be removing mscoree.dll in all dotnet* installs because
it could have been placed by the dotnet40 verb, which would cause
subsequent dotnet* verbs to break dotnet40. The deletion should
instead only happen when the native dotnet is detected. We continue
to remove the registry keys in any case, though, since they are
only relevant to the native DLL.

Closes #2072