Fix `winecmd` when Wine path contains spaces
Fix `winecmd` verb when path to Wine contains a space.

The arguments for `winetricks_shell` were passed as-is in a
space-separated string, meaning a single argument with a space in it
(eg. the Wine path) would break the entire command.

Fix this by single quoting each argument and concatenating them
together. This is a bit tricky in POSIX sh without `printf %q`,
but seems to work fine with the code suggested by ShellCheck for SC3050.