Determine if there is an internet connection first using getent before attempting to download anything.
This method uses getent which is part of glibc, a pretty much hard requirement for any linux system.

Since winetricks is capable of both downloading and installing software as well as performing library
overrides or registry edits, this is useful when you run a script that performs multiple winetricks
with a mixture of all types (such as a lutris install script or protontricks) and don't have an internet
connection. This way anything that does not require an internet connection is not forced to wait
on those that do to fail before continuing.