Allow {WINE,WINESERVER}_BIN env vars
Make it possible to set `WINE_BIN` and `WINESERVER_BIN` using
environment variables outside Winetricks. These affect binary bitness
detection during startup.

Winetricks performs a Wine binary bitness detection to print a warning.
In some environments such as Protontricks and Nix, the Wine executables are
wrappers that call binaries that are located elsewhere. In those
situations we want to provide the path to the binary that's only used
for the check to avoid printing unnecessary warnings.

Closes: #2183