Add French
French language added to Winetricks for graphical interfaces and command line.
I created a new port for translation and the changes made do not affect functionality.

However, there is a small problem with the translation of parameters in kdialog for “title_fr=”. It does not work, whereas in zenity it works perfectly. This will need to be corrected.

I will come back to this later for the DLL titles.

Otherwise, it is ready to be added.

Please note: Use the latest published commit.